Fix element finding on non-English pages.
Data-tooltip that was looked for has different information depending on the language of the user's system, so is unreliable. Have switched to a more reliable, page-agnostic method.

Also added a setting that a user may enable in the code to force auto-play to always be on or off, even when in incognito mode. You will have to edit this value again any time the script is updated though.